refactor(builder): extrahiere findings-Helpers aus single.rs → single/findings.rs
- single.rs → single/mod.rs (Verzeichnis-Modul)
- finding_group_from_normalized, build_location_hints, build_representative_occurrences,
  build_pattern_clusters + private Selector-Helpers → single/findings.rs
- Zugehörige Unit-Tests ziehen mit (3 Tests)
- Ungenutzte Imports aus mod.rs bereinigt
- Keine Logikänderung, alle 608 Unit-Tests + 3 Snapshots grün
